+++ ictonyx/__init__.py
@@ -1,5 +1,5 @@
 # ictonyx/__init__.py
-# v.0.3.2
+# v.0.3.3
 
 """
 Ictonyx: A Machine Learning Framework for Variability and Reproducibility Analysis
@@ -323,7 +323,7 @@ except ImportError:
     pass
 
 # Library version
-__version__ = "0.3.2"
+__version__ = "0.3.3"
 
 
 # Feature availability summary